src/ResearchOS/GUI/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.7.0
mercurial==6.3.2
mysql-connector-python-rf==2.2.2
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pygame==2.2.0
pyparsing==3.0.9
PySide6==6.4.2
PySide6-Addons==6.4.2
PySide6-Essentials==6.4.2
python-dateutil==2.8.2
shiboken6==6.4.2
six==1.16.0
